- did some refactoring on app - tried to integrate formular - not working yet.
