+++ android/app/src/main/java/de/fau/amos/virtualledger/android/MainActivity.java
@@ -3,21 +3,19 @@ package de.fau.amos.virtualledger.android;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
-import android.widget.TextView;
 import android.view.View;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.TextView;
 
 import java.io.IOException;
-import java.util.List;
 
 import javax.inject.Inject;
 
 import de.fau.amos.virtualledger.R;
 import de.fau.amos.virtualledger.android.api.model.StringApiModel;
-import de.fau.amos.virtualledger.android.dagger.Restapi;
-import de.fau.amos.virtualledger.android.dagger.UserCredential;
-import de.fau.amos.virtualledger.android.dagger.component.NetComponent;
+import de.fau.amos.virtualledger.android.api.Restapi;
+import de.fau.amos.virtualledger.android.model.UserCredential;
 import retrofit2.Callback;
 import retrofit2.Response;
 import retrofit2.Retrofit;
@@ -32,24 +30,34 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.testfile);
+
+        // !!! das habe ich geändert - weiß nicht, ob ich das darf?!
+        setContentView(R.layout.registration);
         ((App) getApplication()).getNetComponent().inject(this);
 
-/*        Button button_register = (Button) findViewById(R.id.button_register);
+        Button button_register = (Button) findViewById(R.id.button_register);
 
         button_register.setOnClickListener(
                 new Button.OnClickListener() {
                     public void onClick(View v) {
-                        setContentView(R.layout.register_succ);
+                        register();
                     }
                 }
-        );*/
+        );
+    }
 
+    private void register() {
+
+        String email = ((EditText) findViewById(R.id.Email)).getText().toString();
+        String password = ((EditText) findViewById(R.id.Password)).getText().toString();
+        String firstname = ((EditText) findViewById(R.id.FirstName)).getText().toString();
+        String lastname = ((EditText) findViewById(R.id.LastName)).getText().toString();
+        retrofit2.Call<StringApiModel> responseMessage = retrofit.create(Restapi.class).register(new UserCredential(email, password, firstname, lastname));
+
+        // !!! das funktioniert natürlich jetzt nicht, weil ich auf dem falschen View bin
+        // wie kann ich den View wechseln?!?!
         textView = (TextView) findViewById(R.id.txt);
-        // Create a retrofit call object
-        retrofit2.Call<StringApiModel> responseMessage = retrofit.create(Restapi.class).register(new UserCredential("testasdasdauser55@abc.de", "testpassword", "test", "test2"));
-
-        //Enque the call
+        
         responseMessage.enqueue(new Callback<StringApiModel>() {
             @Override
             public void onResponse(retrofit2.Call<StringApiModel> call, Response<StringApiModel> response) {
