+++ android/app/src/main/java/de/fau/amos/virtualledger/android/api/Restapi.java
@@ -1,13 +1,9 @@
-package de.fau.amos.virtualledger.android.dagger;
-
-import java.util.List;
+package de.fau.amos.virtualledger.android.api;
 
 import de.fau.amos.virtualledger.android.api.model.StringApiModel;
+import de.fau.amos.virtualledger.android.model.UserCredential;
 import retrofit2.Call;
 import retrofit2.http.Body;
-import retrofit2.http.FormUrlEncoded;
-import retrofit2.http.GET;
-import retrofit2.http.Header;
 import retrofit2.http.POST;
 
 /**
